fix: do not erase transferIdAcquirer
Having transferIdAcquirer erased with transfer.push.confirmAcquirer
feels like side effect and unwanted behavior.

If i don't know this ID but i want to confirm the acquirer operation,
i would pass "null" value to it and expect it to be ignored.
If i need to overwrite this value, it is completely different scenario
and should be done with different procedure.

This procedure is not intended to erase this ID.